requirements.txt
boto3==1.4.4
botocore==1.5.56
Django==1.11.1
djangorestframework==3.6.3
docutils==0.13.1
jmespath==0.9.3
pkg-resources==0.0.0
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
